attachment.c file_stream.c header.c io.c mailbox.c mbx_pop.c  	mbx_unix.c mbx_unixscan.c message.c mime.c trans_stream.c  	include/private/io0.h include/public/attribute.h  	include/public/auth.h include/public/body.h  	include/public/header.h include/public/io.h  	include/public/locker.h include/public/mailbox.h  	include/public/message.h
expunge can be call safely, whithout having to close/destroy the mailbox
it will readjust the message_t pointer.

mailbox_t uses stream to now.

+++ mailbox/mailbox.c
@@ -256,7 +256,7 @@ mailbox_deregister (mailbox_t mbox, void *action)
   for (i = 0; i < mbox->event_num; i++)
     {
       event = &(mbox->event[i]);
-      if (event->_action == action)
+      if ((int)event->_action == (int)action)
 	{
 	  event->type = 0;
 	  event->_action = NULL;
